Migrate filenameFromPath over into utils/formatters
