+++ src/shared/formatters.test.js
@@ -54,20 +54,6 @@ describe('formatters', () => {
   });
 });
 
-describe('filenameFromPath', () => {
-  it('returns last portion of path with default delimiter', () => {
-    expect(formatters.filenameFromPath('/home/user/folder/.hidden/My Long Filename.sql')).toEqual(
-      'My Long Filename.sql',
-    );
-  });
-
-  it('returns original filename if no path is included', () => {
-    expect(formatters.filenameFromPath('Just-A-gnarly_filemame(0) DRAFT.v2.docx')).toEqual(
-      'Just-A-gnarly_filemame(0) DRAFT.v2.docx',
-    );
-  });
-});
-
 describe('formatAgeToDays', () => {
   it('returns expected string less than 1 day', () => {
     expect(formatters.formatAgeToDays(0.99)).toEqual('Less than 1 day');
